# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Insect Monitoring API",
    "description": "곤충 모니터링 API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/"
    }
  ],
  "paths": {
    "/account/list": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "사용자 목록",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountListSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/account/get": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "사용자 정보",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountInfoSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralFailedResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/account/register": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "사용자 가입",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountFailedResponse"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "password",
                  "userLevel"
                ],
                "properties": {
                  "userId": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "userLevel": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/account/login": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "로그인",
        "description": "",
        "parameters": [
          {
            "name": "loginId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountFailedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/account/update": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "사용자 수정",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "userLevel"
                ],
                "properties": {
                  "userId": {
                    "type": "integer"
                  },
                  "password": {
                    "type": "string"
                  },
                  "userLevel": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/account/delete": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "사용자 삭제",
        "description": "",
        "responses": {
          "200": {
            "description": "성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneralSuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId"
                ],
                "properties": {
                  "userId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AccountListSuccessResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "array"
          },
          "items": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "object"
              },
              "properties": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "integer"
                      }
                    }
                  },
                  "user_id": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      }
                    }
                  },
                  "password": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "format": {
                        "type": "string",
                        "example": "password"
                      }
                    }
                  },
                  "user_level": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "integer"
                      }
                    }
                  },
                  "last_login_date": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "format": {
                        "type": "string",
                        "example": "date-time"
                      },
                      "nullable": {
                        "type": "boolean",
                        "example": true
                      }
                    }
                  },
                  "created_date": {
                    "type": "object",
                    "properties": {
                      "type": {
                        "type": "string",
                        "example": "string"
                      },
                      "format": {
                        "type": "string",
                        "example": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "AccountInfoSuccessResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  }
                }
              },
              "user_id": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "password": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "password"
                  }
                }
              },
              "user_level": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "integer"
                  }
                }
              },
              "last_login_date": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  },
                  "nullable": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "created_date": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  },
                  "format": {
                    "type": "string",
                    "example": "date-time"
                  }
                }
              }
            }
          }
        }
      },
      "AccountSuccessResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "user": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "object"
                  },
                  "properties": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "integer"
                          },
                          "nullable": {
                            "type": "boolean",
                            "example": true
                          }
                        }
                      },
                      "user_id": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          }
                        }
                      },
                      "password": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "format": {
                            "type": "string",
                            "example": "password"
                          }
                        }
                      },
                      "user_level": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "integer"
                          }
                        }
                      },
                      "last_login_date": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "format": {
                            "type": "string",
                            "example": "date-time"
                          },
                          "nullable": {
                            "type": "boolean",
                            "example": true
                          }
                        }
                      },
                      "created_date": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "string"
                          },
                          "format": {
                            "type": "string",
                            "example": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              },
              "token": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        }
      },
      "AccountFailedResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "user": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "null"
                  }
                }
              },
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        }
      },
...
      "GeneralSuccessResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        }
      },
      "GeneralFailedResponse": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "object"
          },
          "properties": {
            "type": "object",
            "properties": {
              "type": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              },
              "message": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}